requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.43.0
django-el-pagination==3.3.0
django-endless-pagination==2.0
djangorestframework==3.12.2
idna==2.10
joblib==0.17.0
nltk==3.5
numpy==1.19.3
oauthlib==3.1.0
pandas==1.1.4
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.52.0
urllib3==1.26.2
